Backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic[email]==2.8.2

bcrypt==4.1.2
PyJWT==2.9.0
python-multipart==0.0.6

# Database / Supabase
supabase==2.3.3
postgrest==0.14.0
httpx==0.27.0

# ORM Support (optional but required if used)
sqlalchemy==2.0.3

# Utility
python-dotenv==1.0.1
requests==2.32.3
email-validator==2.1.1

# Async + Web
anyio==4.4.0
starlette==0.37.2

# Testing
pytest==8.4.2
pytest-asyncio==0.24.0
pytest-timeout==2.3.1

# CORS (FastAPI adds it, but explicit)
fastapi[all]==0.111.0
